requirements.txt
# Base requirements
torch==2.4.0
torchvision==0.19.0
accelerate>=0.26.0

# Model specific utilities
qwen-vl-utils==0.0.10
transformers-stream-generator==0.0.4
av
pillow>=10.0.0

# API & Serving (New: For FastAPI deployment)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
openai>=1.0.0  # For DeepSeek API (OpenAI-compatible)
bitsandbytes>=0.43.0  # Required for INT4 quantized inference
# Attention notes:
# - SDPA is built into PyTorch 2.x
# - flash-attn is optional and mainly useful on GPUs officially supported by the project

# Install latest transformers from source (Required for Qwen2.5-VL/Vision-R1)
git+https://github.com/huggingface/transformers.git

# Optional but recommended for GPU acceleration
# flash-attn==2.6.1

# For potential future demo usage
gradio==5.4.0
gradio_client==1.4.2
